Truly sync data from remote to local database for piggy bank
Previously, a deleted piggy bank from the web interface is not reflected locally. this resulted in inconsistent data showing. Using kotlin coroutines, data is downloaded from remote datasource and loaded to local datasource. We will then compare the diff of the 2 different datasource and delete data off the local source.